(1/n) rename `DashboardHeader.jsx` to `DashboardHeaderContainer` to avoid collision with `DashboardHeader.tsx`
commit-id:7cdfd86a
